refine code and resolve comments
