+++ mars/oscar/backends/ray/pool.py
@@ -23,14 +23,14 @@ from abc import ABC, abstractmethod
 from enum import Enum
 from typing import List, Optional
 
-from ..message import CreateActorMessage
-from ....serialization.ray import register_ray_serializers
-from ....utils import lazy_import
-from ..config import ActorPoolConfig
-from ..pool import AbstractActorPool, MainActorPoolBase, SubActorPoolBase, create_actor_pool, _register_message_handler
-from ..router import Router
 from .communication import ChannelID, RayServer, RayChannelException
 from .utils import process_address_to_placement, process_placement_to_address, get_placement_group
+from ..config import ActorPoolConfig
+from ..message import CreateActorMessage
+from ..pool import AbstractActorPool, MainActorPoolBase, SubActorPoolBase, create_actor_pool, _register_message_handler
+from ..router import Router
+from ....serialization.ray import register_ray_serializers
+from ....utils import lazy_import
 
 ray = lazy_import('ray')
 logger = logging.getLogger(__name__)
@@ -78,7 +78,7 @@ class RayMainActorPool(MainActorPoolBase):
         actor_handle = ray.remote(RaySubPool).options(
             num_cpus=num_cpus, name=external_address,
             max_concurrency=10000,  # By default, 1000 tasks can be running concurrently.
-            max_restarts=-1,     # Auto restarts by ray
+            max_restarts=-1,  # Auto restarts by ray
             placement_group=pg, placement_group_bundle_index=bundle_index).remote()
         await actor_handle.start.remote(actor_pool_config, process_index)
         return actor_handle
@@ -113,12 +113,17 @@ class RayMainActorPool(MainActorPoolBase):
 
     async def is_sub_pool_alive(self, process: 'ray.actor.ActorHandle'):
         try:
-            await process.health_check.remote()
-            await process.actor_pool.remote('process_index')
+            # try to call the method of sup pool, if success, it's alive.
+            await process.actor_pool.remote('health_check')
             return True
         except Exception:
             logger.info("Detected RaySubPool %s died", process)
             return False
+
+
+class PoolStatus(Enum):
+    HEALTHY = 0
+    UNHEALTHY = 1
 
 
 @_register_message_handler
@@ -133,10 +138,8 @@ class RaySubActorPool(SubActorPoolBase):
         finally:
             self._stopped.set()
 
-
-class PoolStatus(Enum):
-    HEALTHY = 0
-    UNHEALTHY = 1
+    def health_check(self):  # noqa: R0201  # pylint: disable=no-self-use
+        return PoolStatus.HEALTHY
 
 
 class RayPoolBase(ABC):
@@ -174,11 +177,8 @@ class RayPoolBase(ABC):
         """Method for communication based on ray actors"""
         try:
             return await self._ray_server.__on_ray_recv__(channel_id, message)
-        except Exception:   # pragma: no cover
+        except Exception:  # pragma: no cover
             return RayChannelException(*sys.exc_info())
-
-    def health_check(self):  # noqa: R0201  # pylint: disable=no-self-use
-        return PoolStatus.HEALTHY
 
     async def actor_pool(self, attribute, *args, **kwargs):
         attr = getattr(self._actor_pool, attribute)
@@ -194,7 +194,7 @@ class RayPoolBase(ABC):
         try:
             from pytest_cov.embed import cleanup
             cleanup()
-        except ImportError: # pragma: no cover
+        except ImportError:  # pragma: no cover
             pass
